feat(update): build out update package API, use it
