+++ update/daemon.go
@@ -7,22 +7,22 @@ import (
 	"runtime"
 )
 
-// DaemonHelp is a stub for platforms that cannot yet use daemons
-func DaemonHelp() error {
+// daemonHelp is a stub for platforms that cannot yet use daemons
+func daemonHelp() error {
 	return fmt.Errorf("cannot display help for daemon on platform: %s", runtime.GOOS)
 }
 
-// DaemonInstall is a stub for platforms that cannot yet install daemons
-func DaemonInstall() error {
+// daemonInstall is a stub for platforms that cannot yet install daemons
+func daemonInstall() error {
 	return fmt.Errorf("cannot install daemon on platform: %s", runtime.GOOS)
 }
 
-// DaemonUninstall is a stub for platforms that cannot yet uninstall daemons
-func DaemonUninstall() error {
+// daemonUninstall is a stub for platforms that cannot yet uninstall daemons
+func daemonUninstall() error {
 	return fmt.Errorf("cannot uninstall daemon on platform: %s", runtime.GOOS)
 }
 
-// DaemonShow is a stub for platforms that cannot yet show details about daemons
-func DaemonShow() error {
-	return fmt.Errorf("cannot show daemon on platform: %s", runtime.GOOS)
+// daemonShow is a stub for platforms that cannot yet show details about daemons
+func daemonShow() (string, error) {
+	return "", fmt.Errorf("cannot show daemon on platform: %s", runtime.GOOS)
 }
