+++ update/update_test.go
@@ -1,4 +1,4 @@
-package base
+package update
 
 import (
 	"testing"
